requirements.txt
pdfplumber
langchain
faiss-cpu
numpy
requests
python-dotenv
streamlit
unstructured[pdf]
sentence-transformers
langchain-openai


langchain-core
langchain-community
langgraph

unstructured
unstructured-inference

faiss-cpu

huggingface-hub

pydantic
tqdm
